Added SD datalogging and included libraries
Added SD datalogging
Code now logs all sensor values to SD card in a .csv file named with the
date and time at creation. Time is synced with a web server using the
Ethernet shield.

Inlcuded libraries
Added a folder with Arduino libraries used in the project that are not
included in the Arduino core library set.