Added frequency to flowrate conversion, changed Serial to Serial3
Frequency to flowrate conversion
Added equation found using calibration sheet on Dropbox to code.
Previously code was returning the frequency of the flowmeter. Now, the
corresponding flowrate is being returned. Excel document containing graph
and equation is on the Dropbox.

Serial to Serial3
The current transducer data is being sent over RX3/TX3 i.e. Serial3. Now,
the data is being read using Serial3 instead of Serial as before. Serial
is still being used for debugging messages though so that they show up on
a computer that is connected via USB. You'll need to open the Arduino
Serial Monitor or equivalent.