Thermistor equation and Flowmeter equation corrections
The equation that converts thermistor output voltage to temperature has
been updated to match data provided by Greystone. Previous equation was
based on data already on Dropbox. The Excel sheet document is on the
Dropbox.

Flowmeter equation was commented out accidentally. This has been remedied
now. Previously, the flowmeter read function was returning the frequency
instead of the flowrate.